# Don't decode embedded info when creating Id.