old:
```
jvm summary:
Benchmark                     Mode  Cnt  Score    Error  Units
IdBenchmark.generate          avgt    5  0.083 ±  0.001  us/op
IdBenchmark.generateToString  avgt    5  0.178 ±  0.020  us/op
IdBenchmark.stringDecodeToId  avgt    5  0.149 ±  0.055  us/op
```

new:
```
jvm summary:
Benchmark                     Mode  Cnt  Score    Error  Units
IdBenchmark.generate          avgt    5  0.075 ±  0.001  us/op
IdBenchmark.generateToString  avgt    5  0.155 ±  0.008  us/op
IdBenchmark.stringDecodeToId  avgt    5  0.133 ±  0.364  us/op
```